requirements.txt
Django==1.6
South==0.8.4
Unipath==1.0
bootstrap-admin==0.3.0
boto==2.27.0
cloudinary==1.0.16
dj-database-url==0.2.2
dj-static==0.0.5
django-decouple==2.1
django-widget-tweaks==1.3
gunicorn==18.0
mock==1.0.1
nameparser==0.2.8
pystache==0.5.3
pytz==2013.9
static==1.0.2
django-tinymce
cloudinary